Adds a shared helper instance
Apparently, there are problems with multiple helpers/connections when used
concurrently, and they are supposed to be used as singletons. This change adds
a shared instance to a custom Application subclass.

Eventually, it is planned to use DI (Dagger?) to manage dependencies, and the
helper would be a @Singleton managed by the DI container.