Allows the favorite button in test activity to be moved around
I have a Nexus 6 and sometimes (like leaning forward on the dashboard when
riding a stationary bike) the app is very uncomfortable or unusable with one
hand, because the favorite button (which should be the second most often used
action, second only to swiping) is too far away from the fingers. This change
allows the button to be moved around and placed wherever the user wants.