Some minor code changes
Most importantly, if changes the instant when group adding animation is run -
now it is performed in onResume. This fixes the problem when on resume takes
too long and the delay in importTestGroup was not enough.
The refreshing code in onResume is not good either, it might take too long.

Also moves the triggers modifying the index to the Task class.